Add support for per-platform URL templates
Release 4.8-2013-q4 has a slightly different URL for darwin than
for other OSes (since it includes the date and it seems they uploaded
the darwin release a few days later).

To workaround this, add support for checking for a $release-$platform
entry in the URL templates dictionary and, when it exists, give it
priority. Otherwise, use just $release as a key.